requirements.txt
fastapi
uvicorn
pandas
numpy
transformers
pypdf
sqlalchemy
scikit-learn
pyyaml
python-multipart
torch